ocean-plotter-ows/production/requirements.txt
basemap==1.4.1
basemap-data==1.3.2
basemap-data-hires==1.3.2
certifi==2025.1.31
cftime==1.6.4.post1
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
geopandas==1.0.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.8.4
netCDF4==1.7.2
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
shapely==2.0.7
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
xarray==2025.1.2
